+++ src/components/content/Content.form.js
@@ -1,4 +1,12 @@
 import baseEditForm from '../base/Base.form';
+
+import ContentEditDisplay from './editForm/Content.edit.display';
+
 export default function(...extend) {
-  return baseEditForm(...extend);
+  return baseEditForm(...extend, [
+    {
+      key: 'display',
+      components: ContentEditDisplay
+    }
+  ]);
 }